ChatDemo/requirements.txt
streamlit==1.32.0
langchain==0.1.12
openai==1.14.3
faiss-cpu==1.8.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
tiktoken==0.6.0